fix recursion error when getting an invalid action object attribute
This commit updates the ActionContainer __getattr__ method to not use
hasattr to determine if the attribute does not exist. When running
with Python 3, this was causing the method to call itself since
hasattr calls getattr. This update will remove using hasattr and
now use the in keyword. Now if action getattr is called with an
invalid attribute, an attribute error will be rasied.

Closes #33