Add expand and attributes to get and filter methods.
Also added support for wrapping list of dictionaries that are
manageiq objects on Entity reload. This helps with things like when
you expand and get a list of dictionaries back. These are now
wrapped as entities that you can work with rather than dictionaries.

Added the same thing for dictionaries as well.